Step 27: consolidate Next config (TS) + keep web->api mock proxy
