+++ apps/web/next.config.ts
@@ -1,5 +1,18 @@
-import type { NextConfig } from 'next';
+import type { NextConfig } from "next";
 
-const nextConfig: NextConfig = {};
+const nextConfig: NextConfig = {
+  output: "standalone",
+
+  async rewrites() {
+    const upstream = process.env.LOKALTREU_API_UPSTREAM ?? "http://127.0.0.1:3001";
+
+    return [
+      {
+        source: "/api/:path*",
+        destination: `${upstream}/:path*`,
+      },
+    ];
+  },
+};
 
 export default nextConfig;
